Fixed compilation warnings in SM 1.7
Uses new WritePackFunction and ReadPackFunction natives.

This fixes the "warning 237: coercing functions to and from primitives
is unsupported and will be removed in the future" warnings.

Plugins compiled with the 1.7 compiler won't run on 1.6 servers.
If we want to work around this. We need to check for availablity of
those 2 natives at runtime and fall back to the old Function: retagging
method. This would give us the compilation warnings again though..